(weather_tracker) Update dev endpoint config
Update the configuration for dev mode to no longer bind to the loopback
address (`127.0.0.1`) which prevents other machines on the network from
accessing it. Instead, we bind to `0.0.0.0` which *does* allow other
machines on the network to access it, therefore allowing our nerves
device to access the Phoenix server over the LAN.

Thanks to
https://elixirforum.com/t/access-phoenix-dev-in-local-network/43316/2
for the solution here.