(publisher) Create `publisher` project
Create a new project to post data from the sensors attached to our
Nerves device to our Phoenix server.

This application is simply a single GenServer that will pull
periodically pull state from the `sensors` it is passed in on start up.

In addition to creating this project, we pull it into our `sensor_hub`
project's application tree, configuring it as needed.

Setting the LAN IP address in the config directly _probably_ isn't
ideal, but it'll do for now ;p