(weather_tracker) Update fields on `weather_conditions`
Update the fields on the `weather_conditions` table to account for the
data we are getting. These are diffrent than the book because:

- I think for one of the sensors I am keeping additional data the book
ignored.
- For the air quality sensor, I have a newer version of the sensor. The
library that exists seems like it may be providing a calculated
value (`voc_index`) instead of the raw values (I'm guessing that waht
`tvoc_ppb` and maybe `co2_eq_ppm` were).